Move rest of org.gradle.api.internal.file.collections to core-file-services
