+++ subprojects/core-file-services/core-file-services.gradle.kts
@@ -44,4 +44,5 @@ java {
 
 testFixtures {
     from(":core")
+    from(":coreApi")
 }
