+++ subprojects/core-file-services/src/main/java/org/gradle/api/internal/file/collections/SingletonFileSet.java
